+++ thrift4DL/server/http/__init__.py
@@ -3,5 +3,6 @@
 # This software is released under the MIT License.
 # https://opensource.org/licenses/MIT
 
-from .tcp import TModelPoolServer
-from .tcp import VisionHandler
+from .http_server import HTTPServer
+
+